Assumptions are now always rendered
We did not always do this. This leaded  to weird situations.
Suppose you ran `proof (Atom 0 -> Atom 1) [Atom 1]`.
`Atom 0` would be assumed. `Atom 1` could be proven since it was in the theory.
However, in the proof it was not stated that the assumption `Atom 0` was used.
This resulted in the implication introduction rule saying that assumption 0 was used,
while assumption 0 wasn't found in the tree.

When we now want to proof something which is in the theory,
we state that we used all assumptions which were not yet given-up.
This can result in to many assumptions being used.