withAssumption now takes Either Formula Assumption.
When a formula is given, it'll get a number assigned to it.
The assumption counter is then incremented.
When an assumption is given, the assumption counter won't be
incremented. This has the result that both the lhs and the rhs
of a disjunction have the same assumption number when the
or elimination rule is used.