Only one assumption can be given-up per rule.
In 253456f the possiblity to give up multiple assumptions was added.
This was done to make it easier to implement the or elimination.

In 3c6fd85 the possibility to specify the number for an assumption
was added. This was used to have the same number for the lhs and the
rhs assumption in the or elimination.

It is therefore no longer needed to given up more than one assumption
per step.